feat(landing-page): replace old page with new
