--- src/app/landing/layout.tsx
@@ -1,18 +1,0 @@
-import { Footer } from "@/components/footer";
-import { Header } from "@/components/header";
-
-interface LandingLayoutProps {
-  children: React.ReactNode;
-}
-
-export default function LandingLayout({ children }: LandingLayoutProps) {
-  return (
-    <div className="flex flex-col items-center min-h-screen">
-      <Header />
-      <main className="flex-1 flex flex-col items-center w-full">
-        {children}
-      </main>
-      <Footer />
-    </div>
-  );
-}
